[Add] Added AlertDialog [Change] Chage name of icon for clock, folder, locked [Delete] Deleted duplicated calendar and line
