+++ app/src/main/java/com/example/sauexpert/bracelet_indicator/HRVScreen.kt
@@ -79,7 +79,7 @@ fun HRVwithBarChart(
                 TextOfTabData(stringResource(R.string.month_short).toUpperCase(Locale.current)),
                 TextOfTabData(
                     stringResource(R.string.choose).toUpperCase(Locale.current),
-                    painter = painterResource(R.drawable.ic_calendar_icon)
+                    painter = painterResource(R.drawable.ic_calendar)
                 )
             ),
             weight = 0.3f
