+++ app/src/main/java/com/example/sauexpert/bracelet_indicator/PressureScreen.kt
@@ -178,7 +178,7 @@ fun PressureTitle(
                 TextOfTabData(stringResource(R.string.month_short).toUpperCase(Locale.current)),
                 TextOfTabData(
                     stringResource(R.string.choose).toUpperCase(Locale.current),
-                    painter = painterResource(R.drawable.ic_calendar_icon)
+                    painter = painterResource(R.drawable.ic_calendar)
                 )
             ),
             activity = activity,
