+++ app/src/main/java/com/example/sauexpert/bracelet_indicator/PulseScreen.kt
@@ -86,7 +86,7 @@ fun PulsewithBarChart(
                 TextOfTabData(stringResource(R.string.month_short).toUpperCase(Locale.current)),
                 TextOfTabData(
                     stringResource(R.string.choose).toUpperCase(Locale.current),
-                    painter = painterResource(R.drawable.ic_calendar_icon)
+                    painter = painterResource(R.drawable.ic_calendar)
                 )
             ),
             weight = 0.3f
