+++ app/src/main/java/com/example/sauexpert/daily_routine/DailyRoutineScreen.kt
@@ -31,6 +31,7 @@ import com.example.sauexpert.ui.theme.SauExpertTheme
 import com.example.sauexpert.widgets.compose.MainButton
 import com.example.sauexpert.widgets.compose.Toolbars.MainActionToolBar
 import com.example.sauexpert.widgets.compose.buttons.OutlinedMainButton
+import com.example.sauexpert.widgets.compose.pop_up_dailog.RenameDialog
 import kotlinx.coroutines.launch
 
 
@@ -367,89 +368,4 @@ fun BottomSheetContentForDailyRoutine(
     }
 }
 
-
-@Composable
-fun RenameDialog(
-    modifier: Modifier = Modifier,
-    stateForRename: String,
-    onNameChange: (String) -> Unit,
-    isDialogOpen: MutableState<Boolean>
-
-) {
-    if (isDialogOpen.value) {
-        Dialog(onDismissRequest = { isDialogOpen.value = false }) {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center,
-                modifier = Modifier
-                    .background(
-                        color = Gray15,
-                        shape = RoundedCornerShape(14.dp)
-                    )
-                    .padding(top = 18.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.meal_time),
-                    style = MaterialTheme.typography.subtitle2,
-                )
-
-                Spacer(modifier = Modifier.height(15.dp))
-
-                OutlinedTextFieldWithBackground(
-                    textState = stateForRename,
-                    onTextChange = onNameChange,
-                    colorBackground = Color.White,
-//                    textSize = 12.sp,
-                    modifier = modifier
-                        .padding(horizontal = 16.dp)
-                )
-
-
-                Spacer(modifier = Modifier.height(15.dp))
-
-                Divider(
-                    color = Gray30.copy(alpha = 0.19f),
-                    thickness = 1.dp,
-                    modifier = modifier.fillMaxWidth()
-                )
-
-                Row(
-                    modifier = Modifier.height(IntrinsicSize.Min)
-                ) {
-                    OutlinedMainButton(
-                        text = stringResource(id = R.string.cancellation),
-                        onClick = { isDialogOpen.value = false },
-                        enableState = true,
-                        colors = ButtonDefaults.buttonColors(
-                            backgroundColor = Gray15,
-                            contentColor = Blue007AFF,
-                        ),
-                        textColor = Color.Transparent,
-                        modifier = Modifier.weight(0.5f)
-                    )
-
-                    Divider(
-                        color = Gray30.copy(alpha = 0.19f),
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .width(1.dp)
-                    )
-
-                    OutlinedMainButton(
-                        text = stringResource(id = R.string.done),
-                        onClick = { /*TODO*/ },
-                        enableState = true,
-                        colors = ButtonDefaults.buttonColors(
-                            backgroundColor = Gray15,
-                            contentColor = Blue007AFF,
-                        ),
-                        textColor = Color.Transparent,
-                        modifier = Modifier.weight(0.5f)
-                    )
-
-                }
-            }
-        }
-    }
-}
