--- app/src/main/res/drawable/ic_calendar_icon.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="18dp"
-    android:height="17dp"
-    android:viewportWidth="18"
-    android:viewportHeight="17">
-  <path
-      android:pathData="M2.9443,16.8809H15.0469C16.9541,16.8809 17.9473,15.8965 17.9473,14.0156V3.3193C17.9473,1.4385 16.9541,0.4541 15.0469,0.4541H2.9443C1.0459,0.4541 0.0527,1.4297 0.0527,3.3193V14.0156C0.0527,15.8965 1.0459,16.8809 2.9443,16.8809ZM2.9355,15.1318C2.206,15.1318 1.8018,14.7539 1.8018,13.9805V5.8945C1.8018,5.1211 2.206,4.7432 2.9355,4.7432H15.0557C15.7852,4.7432 16.1895,5.1211 16.1895,5.8945V13.9805C16.1895,14.7539 15.7852,15.1318 15.0557,15.1318H2.9355ZM7.3037,7.7754H7.8223C8.1387,7.7754 8.2441,7.6787 8.2441,7.3711V6.8525C8.2441,6.5361 8.1387,6.4394 7.8223,6.4394H7.3037C6.9873,6.4394 6.8818,6.5361 6.8818,6.8525V7.3711C6.8818,7.6787 6.9873,7.7754 7.3037,7.7754ZM10.1777,7.7754H10.6963C11.0039,7.7754 11.1094,7.6787 11.1094,7.3711V6.8525C11.1094,6.5361 11.0039,6.4394 10.6963,6.4394H10.1777C9.8613,6.4394 9.7559,6.5361 9.7559,6.8525V7.3711C9.7559,7.6787 9.8613,7.7754 10.1777,7.7754ZM13.043,7.7754H13.5615C13.8779,7.7754 13.9834,7.6787 13.9834,7.3711V6.8525C13.9834,6.5361 13.8779,6.4394 13.5615,6.4394H13.043C12.7354,6.4394 12.6299,6.5361 12.6299,6.8525V7.3711C12.6299,7.6787 12.7354,7.7754 13.043,7.7754ZM4.4385,10.6055H4.9482C5.2646,10.6055 5.3701,10.5088 5.3701,10.1924V9.6738C5.3701,9.3662 5.2646,9.2695 4.9482,9.2695H4.4385C4.1221,9.2695 4.0166,9.3662 4.0166,9.6738V10.1924C4.0166,10.5088 4.1221,10.6055 4.4385,10.6055ZM7.3037,10.6055H7.8223C8.1387,10.6055 8.2441,10.5088 8.2441,10.1924V9.6738C8.2441,9.3662 8.1387,9.2695 7.8223,9.2695H7.3037C6.9873,9.2695 6.8818,9.3662 6.8818,9.6738V10.1924C6.8818,10.5088 6.9873,10.6055 7.3037,10.6055ZM10.1777,10.6055H10.6963C11.0039,10.6055 11.1094,10.5088 11.1094,10.1924V9.6738C11.1094,9.3662 11.0039,9.2695 10.6963,9.2695H10.1777C9.8613,9.2695 9.7559,9.3662 9.7559,9.6738V10.1924C9.7559,10.5088 9.8613,10.6055 10.1777,10.6055ZM13.043,10.6055H13.5615C13.8779,10.6055 13.9834,10.5088 13.9834,10.1924V9.6738C13.9834,9.3662 13.8779,9.2695 13.5615,9.2695H13.043C12.7354,9.2695 12.6299,9.3662 12.6299,9.6738V10.1924C12.6299,10.5088 12.7354,10.6055 13.043,10.6055ZM4.4385,13.4268H4.9482C5.2646,13.4268 5.3701,13.3301 5.3701,13.0225V12.5039C5.3701,12.1875 5.2646,12.0908 4.9482,12.0908H4.4385C4.1221,12.0908 4.0166,12.1875 4.0166,12.5039V13.0225C4.0166,13.3301 4.1221,13.4268 4.4385,13.4268ZM7.3037,13.4268H7.8223C8.1387,13.4268 8.2441,13.3301 8.2441,13.0225V12.5039C8.2441,12.1875 8.1387,12.0908 7.8223,12.0908H7.3037C6.9873,12.0908 6.8818,12.1875 6.8818,12.5039V13.0225C6.8818,13.3301 6.9873,13.4268 7.3037,13.4268ZM10.1777,13.4268H10.6963C11.0039,13.4268 11.1094,13.3301 11.1094,13.0225V12.5039C11.1094,12.1875 11.0039,12.0908 10.6963,12.0908H10.1777C9.8613,12.0908 9.7559,12.1875 9.7559,12.5039V13.0225C9.7559,13.3301 9.8613,13.4268 10.1777,13.4268Z"
-      android:fillColor="#FF435B"/>
-</vector>
